feat: setup migrations
`typeorm` needs access to a config file at the base level of the project. Either that way or through environment variables. Instead of manually pulling in the config, we're just going to specify at the base level of the project so the connection knows how to be created as well as making it possible to run migrations.