fix: specify return type as array
This ensures the graphql type definition for a Task entity inside of a List entity is `[Task!]!` instead of `Task`. If you run the app and watch the schema definition, you'll see the change. Because the previous definition was specifying that the `task` property on the `List` entity should be an object instead of a list of objects, it was failing basic validation rules. However, I still can't figure out how this worked in previous versions. I'm still not entirely sure what changed on the production machine running the app.

Additionally, I updated several NPM packages and made an NVM directive for using NodeJS v10.15.3.